fix(KUI-1140): refactor
rename removeRoundsOlderThanPreviousYear to roundIsNotOutdated & moving this function to a helper file inside utils folder in server
